requirements.txt
hopsworks[python]
requests
pandas
python-dotenv
scikit-learn